Support PSR-0 composer class loading
