+++ src/Ratchetio.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Rollbar;
+
+
+class Ratchetio extends Rollbar {
+}
