+++ src/SourceFileReader.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Rollbar;
+
+
+class SourceFileReader implements iSourceFileReader {
+
+    public function read_as_array($file_path) { return file($file_path); }
+}
+
