+++ src/iRollbarLogger.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Rollbar;
+
+interface iRollbarLogger {
+    public function log($level, $msg);
+}
+
