+++ src/iSourceFileReader.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Rollbar;
+
+
+interface iSourceFileReader {
+
+    /**
+     * @param string $file_path
+     * @return string[]
+     */
+    public function read_as_array($file_path);
+}
