+++ tests/RollbarTest.php
@@ -131,5 +131,3 @@ class RollbarTest extends PHPUnit_Framework_TestCase {
     }
 
 }
-
-?>
